Added State Saving and Loading
Pulled in dependency on boost serialization to add ability to save
simulation solutions to disk and load those solutions from disk.

To use solution saving, the user must specify the name of the file to save
to in the dictionary passed to SetOptions. The extension of this file
name determines the format of the file. ".txt" saves to a text file (not
human readable), ".xml" saves to an XML file (human readable, good for
debugging and using in other software), and ".bin" saves to a binary
file (fastest saving and loading, boost handles proper byte-ordering).

This is NOT state saving, meaning these files can't be used to continue
a simulation that ended prematurely from wherever it left off.

It is up to the user to call the saving and loading functions at the
appropriate times.

(cherry picked from commit c5c8fa1b8d0903cdb343512f1d27c2da4b7f1f34)